WIP: refactored the Binder verification logic so Failure is much more explicit - rather than inferred from the status of Messages including Pass/Fail/Inconclusive Messages are now reallty diagnostic audit trails rather than the authority.
Rationale: there were issues with Prohibited facets where we had to negate the logic of specific tests, which resulting in the validator logic was becoming too complex - especally when a facet had multiple & optional constraints.

E.g. AttrFacet test: IfcRoot.Name != 'foo' (Prohibited) would never fail

Now the outcome of a test on an entity is independent of the Messages. Calling Fail() on the IdsValidationResult will always result in the failure of the entity. There's no longer any need to invert logic of Failure messages when Prohibiting a facet,

A new enum intro'd for clarity, and to avoid breaking changes with existing interfaces